Level Design
Added props and platforms for character to jump on. Also added a part where the player would be forced to crouch to enter the room under the ramp. Added some lighting props and fire also. The fire is supposed to deter the player from not platforming and running across the ground.